New RL gym environment
introduction of a new way to describe the TVS problem through RL. Now the state vector contains: the equity spot prices, the TVS portfolio spot value and time. This is in agreement with the Hamilton-Jacobi-Bellman problem. The reward function is calculated after each interaction between environment-agent. The reward consists in the difference between V(t+1)-V(t), where V(t) is a proxy of the derivative price.